app/pyproject.toml
[project]
name = "graphrag-demo"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "psycopg2-binary>=2.9.9",
    "pydantic-settings>=2.5.0",
    "sentence-transformers>=3.3.0",
    "anthropic>=0.39.0",
    "openai>=1.55.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.backends._legacy:_Backend"
